requirements.txt
absl-py==0.1.10
chainer==3.3.0
cupy==2.3.0
Cython==0.27.3
h5py==2.7.1
#imageio==2.2.0
ipdb==0.10.3
matplotlib==2.1.2
meshzoo==0.3.1
numpy==1.14.0
opencv-python==3.4.0.12
Pillow==5.0.0
progressbar==2.3
protobuf==3.5.1
PyOpenGL==3.1.0
scikit-image==0.13.1
scipy==1.0.0
torch==0.3.1
torchfile==0.1.0
torchvision==0.2.0
tqdm==4.19.5
visdom==0.1.7
